+++ programs/standalone/examples/TestRunner/insight-devcomp/snippets/Install-Minimal.php
@@ -1,0 +1,17 @@
+<?php
+
+// NOTE: You must have FirePHP Companion installed (http://companion.firephp.org/)
+
+// See FirePHP Companion for result
+
+define('INSIGHT_IPS', '*');
+define('INSIGHT_AUTHKEYS', '*');
+define('INSIGHT_SERVER_PATH', './_insight_.php');
+require_once('FirePHP/Init.php');
+FirePHP::to('controller')->triggerInspect();
+
+
+$inspector = FirePHP::to('request');
+$console = $inspector->console('Test');
+
+$console->log('Hello World');